+++ relayer-cli/src/commands/start_multi.rs
@@ -11,11 +11,10 @@ pub struct StartMultiCmd {}
 impl Runnable for StartMultiCmd {
     fn run(&self) {
         let config = app_config();
-
         let supervisor = Supervisor::spawn(config.clone()).expect("failed to spawn supervisor");
         match supervisor.run() {
             Ok(()) => Output::success_msg("done").exit(),
-            Err(e) => Output::error(format!("{}", e)).exit(),
+            Err(e) => Output::error(e).exit(),
         }
     }
 }